requirements.txt
appdirs==1.4.3
cx-Freeze==5.0.1
future==0.16.0
olefile==0.44
packaging==16.8
Pillow==4.1.0
psutil==5.2.2
PyAutoGUI==0.9.36
PyMsgBox==1.0.6
pyparsing==2.2.0
pywin32==301
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyScreeze==0.1.9
PyTweening==1.0.3
pywin32==301
six==1.10.0
